requirements.txt
streamlit>=1.28
langchain==0.0.347
langchain-core==0.0.11
langchainhub==0.1.14
openai>=1.2
tiktoken==0.5.2
faiss-cpu==1.7.4
python-dotenv==1.0.0